+++ api/python/quilt3/_graphql_client/client.py
@@ -21,27 +21,11 @@ from .bucket_tabulator_tables_list import (
     BucketTabulatorTablesList,
     BucketTabulatorTablesListBucketConfig,
 )
-from .enums import SearchResultOrder
-from .input_types import (
-    PackagesSearchFilter,
-    PackageUserMetaPredicate,
-    UserInput,
-)
+from .input_types import UserInput
 from .roles_list import (
     RolesList,
     RolesListRolesManagedRole,
     RolesListRolesUnmanagedRole,
-)
-from .search_more_packages import (
-    SearchMorePackages,
-    SearchMorePackagesSearchMorePackagesInvalidInput,
-    SearchMorePackagesSearchMorePackagesPackagesSearchResultSetPage,
-)
-from .search_packages import (
-    SearchPackages,
-    SearchPackagesSearchPackagesEmptySearchResultSet,
-    SearchPackagesSearchPackagesInvalidInput,
-    SearchPackagesSearchPackagesPackagesSearchResultSet,
 )
 from .sso_config_get import SsoConfigGet, SsoConfigGetAdminSsoConfig
 from .sso_config_set import (
@@ -81,9 +65,7 @@ def gql(q: str) -> str:
 
 
 class Client(BaseClient):
-    def roles_list(
-        self, **kwargs: Any
-    ) -> List[Union[RolesListRolesUnmanagedRole, RolesListRolesManagedRole]]:
+    def roles_list(self, **kwargs: Any) -> List[Union[RolesListRolesUnmanagedRole, RolesListRolesManagedRole]]:
         query = gql(
             """
             query rolesList {
@@ -112,9 +94,7 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {}
-        response = self.execute(
-            query=query, operation_name="rolesList", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="rolesList", variables=variables, **kwargs)
         data = self.get_data(response)
         return RolesList.model_validate(data).roles
 
@@ -168,9 +148,7 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"name": name}
-        response = self.execute(
-            query=query, operation_name="usersGet", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersGet", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersGet.model_validate(data).admin.user.get
 
@@ -224,13 +202,13 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {}
-        response = self.execute(
-            query=query, operation_name="usersList", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersList", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersList.model_validate(data).admin.user.list
 
-    def users_create(self, input: UserInput, **kwargs: Any) -> Union[
+    def users_create(
+        self, input: UserInput, **kwargs: Any
+    ) -> Union[
         UsersCreateAdminUserCreateUser,
         UsersCreateAdminUserCreateInvalidInput,
         UsersCreateAdminUserCreateOperationError,
@@ -306,15 +284,11 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"input": input}
-        response = self.execute(
-            query=query, operation_name="usersCreate", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersCreate", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersCreate.model_validate(data).admin.user.create
 
-    def users_delete(
-        self, name: str, **kwargs: Any
-    ) -> Optional[UsersDeleteAdminUserMutate]:
+    def users_delete(self, name: str, **kwargs: Any) -> Optional[UsersDeleteAdminUserMutate]:
         query = gql(
             """
             mutation usersDelete($name: String!) {
@@ -348,15 +322,11 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"name": name}
-        response = self.execute(
-            query=query, operation_name="usersDelete", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersDelete", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersDelete.model_validate(data).admin.user.mutate
 
-    def users_set_email(
-        self, email: str, name: str, **kwargs: Any
-    ) -> Optional[UsersSetEmailAdminUserMutate]:
+    def users_set_email(self, email: str, name: str, **kwargs: Any) -> Optional[UsersSetEmailAdminUserMutate]:
         query = gql(
             """
             mutation usersSetEmail($email: String!, $name: String!) {
@@ -430,15 +400,11 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"email": email, "name": name}
-        response = self.execute(
-            query=query, operation_name="usersSetEmail", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersSetEmail", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersSetEmail.model_validate(data).admin.user.mutate
 
-    def users_set_admin(
-        self, name: str, admin: bool, **kwargs: Any
-    ) -> Optional[UsersSetAdminAdminUserMutate]:
+    def users_set_admin(self, name: str, admin: bool, **kwargs: Any) -> Optional[UsersSetAdminAdminUserMutate]:
         query = gql(
             """
             mutation usersSetAdmin($name: String!, $admin: Boolean!) {
@@ -512,15 +478,11 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"name": name, "admin": admin}
-        response = self.execute(
-            query=query, operation_name="usersSetAdmin", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersSetAdmin", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersSetAdmin.model_validate(data).admin.user.mutate
 
-    def users_set_active(
-        self, active: bool, name: str, **kwargs: Any
-    ) -> Optional[UsersSetActiveAdminUserMutate]:
+    def users_set_active(self, active: bool, name: str, **kwargs: Any) -> Optional[UsersSetActiveAdminUserMutate]:
         query = gql(
             """
             mutation usersSetActive($active: Boolean!, $name: String!) {
@@ -594,15 +556,11 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"active": active, "name": name}
-        response = self.execute(
-            query=query, operation_name="usersSetActive", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersSetActive", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersSetActive.model_validate(data).admin.user.mutate
 
-    def users_reset_password(
-        self, name: str, **kwargs: Any
-    ) -> Optional[UsersResetPasswordAdminUserMutate]:
+    def users_reset_password(self, name: str, **kwargs: Any) -> Optional[UsersResetPasswordAdminUserMutate]:
         query = gql(
             """
             mutation usersResetPassword($name: String!) {
@@ -636,12 +594,7 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"name": name}
-        response = self.execute(
-            query=query,
-            operation_name="usersResetPassword",
-            variables=variables,
-            **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersResetPassword", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersResetPassword.model_validate(data).admin.user.mutate
 
@@ -651,7 +604,7 @@ class Client(BaseClient):
         role: str,
         append: bool,
         extra_roles: Union[Optional[List[str]], UnsetType] = UNSET,
-        **kwargs: Any
+        **kwargs: Any,
     ) -> Optional[UsersSetRoleAdminUserMutate]:
         query = gql(
             """
@@ -727,15 +680,11 @@ class Client(BaseClient):
             "extraRoles": extra_roles,
             "append": append,
         }
-        response = self.execute(
-            query=query, operation_name="usersSetRole", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersSetRole", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersSetRole.model_validate(data).admin.user.mutate
 
-    def users_add_roles(
-        self, name: str, roles: List[str], **kwargs: Any
-    ) -> Optional[UsersAddRolesAdminUserMutate]:
+    def users_add_roles(self, name: str, roles: List[str], **kwargs: Any) -> Optional[UsersAddRolesAdminUserMutate]:
         query = gql(
             """
             mutation usersAddRoles($name: String!, $roles: [String!]!) {
@@ -805,18 +754,12 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"name": name, "roles": roles}
-        response = self.execute(
-            query=query, operation_name="usersAddRoles", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersAddRoles", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersAddRoles.model_validate(data).admin.user.mutate
 
     def users_remove_roles(
-        self,
-        name: str,
-        roles: List[str],
-        fallback: Union[Optional[str], UnsetType] = UNSET,
-        **kwargs: Any
+        self, name: str, roles: List[str], fallback: Union[Optional[str], UnsetType] = UNSET, **kwargs: Any
     ) -> Optional[UsersRemoveRolesAdminUserMutate]:
         query = gql(
             """
@@ -891,12 +834,7 @@ class Client(BaseClient):
             "roles": roles,
             "fallback": fallback,
         }
-        response = self.execute(
-            query=query,
-            operation_name="usersRemoveRoles",
-            variables=variables,
-            **kwargs
-        )
+        response = self.execute(query=query, operation_name="usersRemoveRoles", variables=variables, **kwargs)
         data = self.get_data(response)
         return UsersRemoveRoles.model_validate(data).admin.user.mutate
 
@@ -956,9 +894,7 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {}
-        response = self.execute(
-            query=query, operation_name="ssoConfigGet", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="ssoConfigGet", variables=variables, **kwargs)
         data = self.get_data(response)
         return SsoConfigGet.model_validate(data).admin.sso_config
 
@@ -1044,9 +980,7 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"config": config}
-        response = self.execute(
-            query=query, operation_name="ssoConfigSet", variables=variables, **kwargs
-        )
+        response = self.execute(query=query, operation_name="ssoConfigSet", variables=variables, **kwargs)
         data = self.get_data(response)
         return SsoConfigSet.model_validate(data).admin.set_sso_config
 
@@ -1066,21 +1000,12 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"name": name}
-        response = self.execute(
-            query=query,
-            operation_name="bucketTabulatorTablesList",
-            variables=variables,
-            **kwargs
-        )
+        response = self.execute(query=query, operation_name="bucketTabulatorTablesList", variables=variables, **kwargs)
         data = self.get_data(response)
         return BucketTabulatorTablesList.model_validate(data).bucket_config
 
     def bucket_tabulator_table_set(
-        self,
-        bucket_name: str,
-        table_name: str,
-        config: Union[Optional[str], UnsetType] = UNSET,
-        **kwargs: Any
+        self, bucket_name: str, table_name: str, config: Union[Optional[str], UnsetType] = UNSET, **kwargs: Any
     ) -> Union[
         BucketTabulatorTableSetAdminBucketSetTabulatorTableBucketConfig,
         BucketTabulatorTableSetAdminBucketSetTabulatorTableInvalidInput,
@@ -1123,16 +1048,9 @@ class Client(BaseClient):
             "tableName": table_name,
             "config": config,
         }
-        response = self.execute(
-            query=query,
-            operation_name="bucketTabulatorTableSet",
-            variables=variables,
-            **kwargs
-        )
+        response = self.execute(query=query, operation_name="bucketTabulatorTableSet", variables=variables, **kwargs)
         data = self.get_data(response)
-        return BucketTabulatorTableSet.model_validate(
-            data
-        ).admin.bucket_set_tabulator_table
+        return BucketTabulatorTableSet.model_validate(data).admin.bucket_set_tabulator_table
 
     def bucket_tabulator_table_rename(
         self, bucket_name: str, table_name: str, new_table_name: str, **kwargs: Any
@@ -1179,15 +1097,10 @@ class Client(BaseClient):
             "newTableName": new_table_name,
         }
         response = self.execute(
-            query=query,
-            operation_name="bucketTabulatorTableRename",
-            variables=variables,
-            **kwargs
+            query=query, operation_name="bucketTabulatorTableRename", variables=variables, **kwargs
         )
         data = self.get_data(response)
-        return BucketTabulatorTableRename.model_validate(
-            data
-        ).admin.bucket_rename_tabulator_table
+        return BucketTabulatorTableRename.model_validate(data).admin.bucket_rename_tabulator_table
 
     def tabulator_get_open_query(self, **kwargs: Any) -> bool:
         query = gql(
@@ -1200,12 +1113,7 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {}
-        response = self.execute(
-            query=query,
-            operation_name="tabulatorGetOpenQuery",
-            variables=variables,
-            **kwargs
-        )
+        response = self.execute(query=query, operation_name="tabulatorGetOpenQuery", variables=variables, **kwargs)
         data = self.get_data(response)
         return TabulatorGetOpenQuery.model_validate(data).admin.tabulator_open_query
 
@@ -1222,157 +1130,6 @@ class Client(BaseClient):
             """
         )
         variables: Dict[str, object] = {"enabled": enabled}
-        response = self.execute(
-            query=query,
-            operation_name="tabulatorSetOpenQuery",
-            variables=variables,
-            **kwargs
-        )
+        response = self.execute(query=query, operation_name="tabulatorSetOpenQuery", variables=variables, **kwargs)
         data = self.get_data(response)
-        return TabulatorSetOpenQuery.model_validate(
-            data
-        ).admin.set_tabulator_open_query.tabulator_open_query
-
-    def search_packages(
-        self,
-        buckets: Union[Optional[List[str]], UnsetType] = UNSET,
-        search_string: Union[Optional[str], UnsetType] = UNSET,
-        filter: Union[Optional[PackagesSearchFilter], UnsetType] = UNSET,
-        user_meta_filters: Union[
-            Optional[List[PackageUserMetaPredicate]], UnsetType
-        ] = UNSET,
-        latest_only: Union[Optional[bool], UnsetType] = UNSET,
-        size: Union[Optional[int], UnsetType] = UNSET,
-        order: Union[Optional[SearchResultOrder], UnsetType] = UNSET,
-        **kwargs: Any
-    ) -> Union[
-        SearchPackagesSearchPackagesPackagesSearchResultSet,
-        SearchPackagesSearchPackagesEmptySearchResultSet,
-        SearchPackagesSearchPackagesInvalidInput,
-    ]:
-        query = gql(
-            """
-            query searchPackages($buckets: [String!], $searchString: String, $filter: PackagesSearchFilter, $userMetaFilters: [PackageUserMetaPredicate!], $latestOnly: Boolean = false, $size: Int = 30, $order: SearchResultOrder = BEST_MATCH) {
-              searchPackages(
-                buckets: $buckets
-                searchString: $searchString
-                filter: $filter
-                userMetaFilters: $userMetaFilters
-                latestOnly: $latestOnly
-              ) {
-                __typename
-                ... on PackagesSearchResultSet {
-                  ...PackagesSearchResultSetSelection
-                }
-                ... on EmptySearchResultSet {
-                  _
-                }
-                ... on InvalidInput {
-                  ...InvalidInputSelection
-                }
-              }
-            }
-
-            fragment InvalidInputSelection on InvalidInput {
-              errors {
-                path
-                message
-                name
-                context
-              }
-            }
-
-            fragment PackagesSearchResultSetSelection on PackagesSearchResultSet {
-              total
-              firstPage(size: $size, order: $order) {
-                cursor
-                hits {
-                  ...SearchHitPackageSelection
-                }
-              }
-            }
-
-            fragment SearchHitPackageSelection on SearchHitPackage {
-              id
-              score
-              bucket
-              name
-              modified
-              size
-              hash
-              comment
-            }
-            """
-        )
-        variables: Dict[str, object] = {
-            "buckets": buckets,
-            "searchString": search_string,
-            "filter": filter,
-            "userMetaFilters": user_meta_filters,
-            "latestOnly": latest_only,
-            "size": size,
-            "order": order,
-        }
-        response = self.execute(
-            query=query, operation_name="searchPackages", variables=variables, **kwargs
-        )
-        data = self.get_data(response)
-        return SearchPackages.model_validate(data).search_packages
-
-    def search_more_packages(
-        self, after: str, size: Union[Optional[int], UnsetType] = UNSET, **kwargs: Any
-    ) -> Union[
-        SearchMorePackagesSearchMorePackagesPackagesSearchResultSetPage,
-        SearchMorePackagesSearchMorePackagesInvalidInput,
-    ]:
-        query = gql(
-            """
-            query searchMorePackages($after: String!, $size: Int = 30) {
-              searchMorePackages(after: $after, size: $size) {
-                __typename
-                ... on PackagesSearchResultSetPage {
-                  ...PackagesSearchResultSetPageSelection
-                }
-                ... on InvalidInput {
-                  ...InvalidInputSelection
-                }
-              }
-            }
-
-            fragment InvalidInputSelection on InvalidInput {
-              errors {
-                path
-                message
-                name
-                context
-              }
-            }
-
-            fragment PackagesSearchResultSetPageSelection on PackagesSearchResultSetPage {
-              cursor
-              hits {
-                ...SearchHitPackageSelection
-              }
-            }
-
-            fragment SearchHitPackageSelection on SearchHitPackage {
-              id
-              score
-              bucket
-              name
-              modified
-              size
-              hash
-              comment
-            }
-            """
-        )
-        variables: Dict[str, object] = {"after": after, "size": size}
-        response = self.execute(
-            query=query,
-            operation_name="searchMorePackages",
-            variables=variables,
-            **kwargs
-        )
-        data = self.get_data(response)
-        return SearchMorePackages.model_validate(data).search_more_packages
+        return TabulatorSetOpenQuery.model_validate(data).admin.set_tabulator_open_query.tabulator_open_query