+++ api/python/tests/test_admin_api.py
@@ -35,7 +35,10 @@ from .fixtures.graphql_schema_fragments import (
     validate_user_response,
 )
 # Import GraphQL fixtures
-from .graphql_fixtures import graphql_router
+from .graphql_fixtures import (  # pylint: disable=unused-import
+    graphql_router,
+    mock_admin_client,
+)
 
 USER_MUTATION_ERRORS = (
     (INVALID_INPUT_ERROR, admin.Quilt3AdminError),
@@ -53,7 +56,9 @@ def _as_dataclass_kwargs(data: dict) -> dict:
         "typename__" if k == "__typename" else _camel_to_snake(k): (
             _as_dataclass_kwargs(v)
             if isinstance(v, dict)
-            else [_as_dataclass_kwargs(x) for x in v] if isinstance(v, list) else v
+            else [_as_dataclass_kwargs(x) for x in v]
+            if isinstance(v, list)
+            else v
         )
         for k, v in data.items()
     }