+++ docs/CHANGELOG.md
@@ -16,6 +16,12 @@ Entries inside each section should be ordered by type:
 
 # Changelog
 
+## 7.0.0 - 2025-08-25
+
+### Python API
+
+* [Changed] **BREAKING:** Change the default `selector_fn` for `Package.push`. Now, files in the same bucket as the destination registry are not copied by default. This change was requested by users who were inadvertently creating extra copies of files in packages that were created from files in S3. ([#4464](https://github.com/quiltdata/quilt/pull/4464))
+
 ## 6.3.1 - 2025-03-10
 
 ### Python API